[docs] Add secrets API information to design doc
This commit adds the secrets API to the design document. Included
are the following endpoints:

  * POST /secrets
  * GET /secrets
  * GET /secrets/{{secret_id}}
  * DELETE /secrets/{{secret_id}}

In addition, the "Provided Utility Document Kinds" was expounded
on with information related to secret types in Barbican.

The policy actions that are included for these API endpoints
are Barbican policy actions. Note that additional endpoints
might be enforced by Barbican along the way.

Change-Id: Icdb83a36853d5791bf15e275152472137d5156cf