Revision diffing API.
This commit implements revision diffing API and the
required back-end logic needed to realize it.

This commit includes unit tests and unskips as many
of the functional tests related to revision diffing as
possible. This includes all functional tests except for
those that have to do with document deletion, which
will be implemented in a separate commit.

Change-Id: I0419ee9b8cf3fca6fe75818615d2338dc00b1003