Update patch set 3
Patch Set 3: Code-Review+1

Using lax inequality and upper constraint for dev and then pin to your heart's content for test and prod would be a more sensible approach then to pin all the requirements during development - it is as if we want to keep updating these pinned dependencies to pick up incremental fixes and update.  I am not objecting to frequent update to pins if you don't care for the maintainence.

Patch-set: 3
Label: Code-Review=+1