Clean up Deckhand 405/404 error handling
If a method is not allowed, Deckhand currently uses convoluted logic
to throw a 405. Instead, Deckhand should rely on falcon's native
logic which already does that.

This PS relies on falcon logic to throw a 404/405 when an endpoint
is invalid or if a HTTP verb is not support for an endpoint
(respectively).

This PS also adds a VersionsResource to Deckhand to allow information
about the API versions to be returned. This implementation is modeled
after Shipyard [0].

[0] https://github.com/att-comdev/shipyard/blob/c84e91bad116a1d0d4b457665a11bab72aff4264/shipyard_airflow/control/api.py#L32

Change-Id: I805801892dbe1a6bca4d0ba5d445fee554c256f8