Update policy and validation design documentation
Even though the design documentation is eventually going to be
subsumed by tox -e docs (automatic documentation generation)
to facilitate policy documentation for example, the policy
documentation in the design.md file is out of date and
should be corrected until then.

Also correct confusing back-and-forth usage of "success"
and "succeeded" or "failed" and "failure" for validation
statuses.

Finally, fix a small typo in in-code policy docs.

Change-Id: I461053ef520e64ce6c3576b1201b2aacc52f6207