Fix up tags attribute in revisions API
Currently the GET /revisions endpoint returns the tags associated
with reach revision like:

    tags:
      foo:
        name: foo

Which is strange, because the name of the tag is recursively referenced
also as `name: foo`. Thus, this PS re-formats this response body to
be more sensible:

    tags:
      - name: foo
        data: {}

Where `data: {}` indicates that there is no data associated with the tag.

This PS also makes necessary documentation and functional test updates.

Change-Id: I3156a539da11c40fee403806b3bd72b62d5b555d