Update patch set 2
Patch Set 2:

I think shipyard will need a matching update for this change, as it will not work for 
if "committed" in tags:
    # do things

Curious why not  just drop the name element? 

tags:
    foo:
        data: {}

allowing for mapped lookup of tags instead of needing to iterate them and dig out the name?

Patch-set: 2
Label: Workflow=0