Add health resource for ucp-integration API convention
It is a UCP API convention to include a health resource in order
for other components to access and validate Deckhand's health status [0].

As such, this PS accomplishes that goal. Also add uwsgi.ini file to
instantiate the webserver using a more complex configuration that
can be overriden more easily.

[0] https://github.com/att-comdev/ucp-integration/blob/master/docs/api-conventions.md#health-check-api

Change-Id: Ice24cec6d0b98c16af62d9436925083d4092a032