Add expected errors decorator for more resiliency
To provide better resiliency, this PS adds a decorator called
``expected_errors`` which can be applied to controller HTTP
verb functions like on_get. A list of HTTP codes can be specified.
If one of those is detected, the error is raised as an
appropriate falcon variety (e.g. 400 => falcon.BadRequest).

If the exception is already a falcon-based exception type,
it is raised as normal.

Otherwise the exception is treated as unexpected and is raised
as a ``falcon.HTTPInternalServerError``.

Change-Id: If7f8c3bf6b6ada7697611a0bef7bf8f635fc0b7f