Unit tests for health/versions controller
This PS adds unit tests for health and versions controller.
In doing so, it was discovered that there was a bug related to
unconditionally adding a response header of content-type: x-yaml
to the response, because this is not allowed when the response
is a 204. This PS also fixes that bug.

Change-Id: If97cb3bfaca8f71416c2664f3af9cbf96329c3b4