Default allow anonymous access
Switches the chart to by default enable anonymous access to
deckhand.

This causes a couple of side effects, the desirable
one being that the api/v1.0/health endpoint and the /versions
endpoint can be accessed without first authenticating.

The second side effect is somewhat less desired, as it means that
an unauthenticated user accessing a resource that requires
authorization will instead be returned a 403 forbidden when in
fact they are 401 anauthorized (unauthenticated).

This could be remedied and avoided if the check for authentication
were moved into the authorization enforcer decorator
(policy.authorize) to defer the check for authentication until
after it has been determined that the endpoint is indeed guarded by
an auth policy.

Currently the behavior to reject requests as unauthenticated is done
in middleware (middleware.ContextMiddleware), which causes the
enforcement to occur without regard for the endpoint.

Change-Id: I369087430b403355ce551384b8d798465738b684