Update patch set 1
Patch Set 1:

Filipe - thanks for challenging this particular implementation - I think it speaks to a larger decision that needs to be made.

In terms of a security concern, there are two that I recognize in this case:

FIrst, that in the future, a mistaken lack of annotation/decoration would leave an endpoint unsecured. 

Second that having _any_ unsecured endpoints has at least the possibility of opening an attack vector for DOS, and potentially more if the service itself exposes information or provides an injection attack vector. 


If we want to change the basic api/v1.0/health check to include authentication, that needs to be something we document and change across the components in a consistent way. The approach that was being taken (and evidently not well enough documented) was that the basic health check was openly accessible, while the extended health check (when implemented) was not, and would require not just authentication but also authorization. 

I did a little searching, and the documentation shows openstack components (swift) placing its healthcheck outside the boundary of authtoken see [0] 
Also the "more standard" oslo healthcheck [1] shows 'place health check at the beginning of the pipeline.'
It appears that the oslo healthcheck then provides a means to hook methods in the back-end to do specific checks.
Obviously re-arranging the order of the paste pipeline can allow for authentication to occur before the healthcheck.

I'm not proposing we use the oslo healthcheck mechanism as it seems over-engineered for the purposes, but the take-away that I'm getting is that it is not unnatural to have an unauthenticated healthcheck.


My reasoning for the approach taken thus far is:

* The falcon framework we're using make it easy to add endpoints in the code itself, and in a straightforward way allows for the component hosting the healthcheck to determine what to check.  Initially this being only a response indicates to the consumer that at least the uwsgi startup of the application didn't error out and enter into a running but non-responsive mode, but could do more controlled checks in the future.

*  In support of concepts like preflight checks and monitoring, an unauthenticated <basic> health check provides ease of use by multiple kinds of tooling without overt negotiation of a auth token.


So... I'm taking some action items here:

1) Need to engage the product owners to determine if there is a desire to have the basic healthcheck be accessible or non-accessible to an unauthenticated client.

1a) clarify this in the ucp-integration api-conventions

1b) Update the UCP components to match the decided approach - this change will need to be modified to match, and I believe the commit message hints at some of the ways this might be done if the unauthenticated health check stands.

2) Need to update each of the UCP components to make the versions api be protected by a policy, so that authorization is enforced there as well.


[0] https://docs.openstack.org/mitaka/config-reference/image-service/glance-api-paste.ini.html
[1] https://docs.openstack.org/keystone/pike/configuration.html

Side note: I have no designs of making the versions endpoint unauthenticated in particular, but it seems to me that for ANY request that is required to be authenticated, it should also be authorized by a policy that is then open to "all authenticated users", such that it could be changed via configuration to be more stringent.

Patch-set: 1