Update patch set 1
Patch Set 1:

> think it speaks to a larger decision that needs to be made.

I think this is the crux of it. While a pattern was established for using Keystone-backed AAA, the edge-cases were not ironed out. Health checks are a good example. They should not require authentication for two reasons 1) Kubernetes is unlikely to support getting a Keystone token in order for it to check the health of services and pods and 2) the more hurdles you have in a health check, the less direct information a failed health check gives you. If you require 3rd party auth, then a failed health check only says that "3rd party auth is down" OR "I don't have the correct credentials" OR "the target service is not responding". Optimally, the last one is the only information you want a failed health check to represent.

As for the correct implementation of anonymous requests, I think Keystone allows for this with the 'delayed auth' model. UCP components will just need to put the adequate policies/middleware in place to utilize it correctly.

Bryan's point about a missing decorator opening a handler is well received - it seems we probably want to enhance the Falcon middleware / policy decorator system a bit to force a request handler resource to explicitly declare support for anonymous requests. This would allow the Falcon middleware to reject a anonymous request even if a policy decorator was not in place.

Patch-set: 1
Label: Workflow=0