Update patch set 1
Patch Set 1:

> Just to be more complete with the preceding comment:
 > * My examples pointing to the OpenStack implementations of
 > healthcheck are simply meant as evidence, not as proof of goodness.
 > * the oslo middleware referenced is in various stages of
 > implementation per the blueprint, so perhaps not completely vetted
 > by their community (I don't know)
 > * The oslo middleware for healthcheck clearly states: "It’s
 > recommanded to block this url or to randomize the url used for
 > healthcheck, because this feature it’s cleary dedicated for tools
 > used by a deployer like load balancer."  So those same concerns
 > must be considered here as well.

Thanks for the detailed explanation. If we want to do an unauthed health check though, it should just apply to the health check endpoint via a more sophisticated mechanism. However, enabling unauthenticated requests for all endpoints isn't the right approach and I think that CSO will probably rightfully oppose this approach.

I'd actually like to remove this configuration option and simply use mocking via fixtures for unit and functional tests (gabbi can use fixtures, too). Reason for the removal is that it opens the door for insecure configuration.

That being said, some middleware hook like oslo.healthcheck seems sensible. If it's not viable for whatever reason, a simple try/catch that squashes the exception for the health check endpoint is a worst-case solution. This proposed solution is too insecure however.

Patch-set: 1