Update patch set 19
Patch Set 19:

> I'd like to make sure Mark has an opportunity to review

There are a couple other edge cases that aren't currently covered/validated:
  - If a document specifies a parentSelector but the parent can't be found is this a validation error?
  - If permissions prevent a secret document from being retrieved causing substitution to raise a 404 when looking for a secret document is this OK?
  - Do we want to disable layering at all? I think that one possibility is to set the `layerOrder` in the layeringPolicy to be an empty list, causing layering to be ignored.

Other main question that isn't covered: Do we want to accumulate substitution/layering issues and return them all to the user, or just individually raise an error as soon as it happens?

Most of that I feel like is out of scope for this PS and should be approached in follow-ups.

Patch-set: 19
Label: Code-Review=0