Clean up document validation module
This PS cleans up the document validation module
to make it more readable and also to fix the
previous FIXME that was included inside the module:

    Remove this [*] once all Deckhand tests have been
    refactored to account for dynamic schema registeration via
    `DataSchema` documents. Otherwise most tests will fail.

* Refers to the fact that a workaround was being used to
  get around the fact that unregistered document schemas
  could pass validation.

This behavior has been changed so that, given a schema
like foo/bar/v1, it must be registered via a DataSchema
document with `metadata.name`=foo/bar/v1 for the
schema to be detected by Deckhand and then validated
against the appropriate documents.

This PS reworks some tests to account for this new
behavior.

Change-Id: Ie75742b984b7ad392cb41decc203d42842050c80