Test: add unusual documents to functional testing
Unusual documents are documents with different data
types for the data field. The data types include:
object, array, string and integer.

This PS makes necessary ORM model and schema
changes needed to support the different data types.

The ORM data type for the data column has been changed
to PickleType, which "builds upon the Binary type to apply
Python’s pickle.dumps() to incoming objects, and pickle.loads()
on the way out, allowing any pickleable Python object to be
stored as a serialized binary field" -- thereby trivializing
the serialization of different types [0]. The Binary
type itself is uses BLOB for MySQL and BYTEA for PostgreSQL.

[0] http://docs.sqlalchemy.org/en/latest/core/type_basics.html#sqlalchemy.types.PickleType

Change-Id: I53694d56bef71adacb5eb79162678be73acb4ad8