Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

I don't think we should use pickle for this.  I don't see why we're avoiding the postgres data type here, when we know it's available.  There's no value in being agnostic for an application like this.

The specific handling of strings is pretty awkward, but presumably that can be changed later when we have functional tests that integrate with an actual encryption backend.

Patch-set: 2
Label: Code-Review=-1