Update DeckHand Chart - Multi-Threads/Workers
This p.s. will allow multi-thread/worker parameters to be
configurable in the Deckhand chart so that the values can
be injected into the pod environment. This is a follow up
to the comments made in [0].

Note also that we will need multiple workers in order to
handle concurrent requests from Armada and DryDock to DeckHand
for the rendered document. Multi-threads with single worker
did not work as expected. Test results from our lab environment
suggests that 4 single-threaded workers will be sufficient
for our purpose. Hence we will use that as default override
values for now.

[0] https://review.gerrithub.io/#/c/393679/

Change-Id: I228713ec7b2ec305cbc2c761bc77125ea98e7dfa