functional tests: Dump logs to stdout/stderr
This PS creates a logging.conf file to be used inside the Deckhand
container for functional tests. The logging.conf file overrides
logging options to dump all messages to stdout and stderr, which
is the common practice for container logging. This file is
referenced in turn by the Deckhand configuration file.

Now it should be possible to actually see debug and error messages
in the CICD gate following functional test execution failure.

Currently, failures are resulting in completely useless output
like this:

  === Deckhand Server Log ===
  + cat deckhand.log
  cat: deckhand.log: No such file or directory
  + cleanup
  ...

Which in other words makes debugging using the Jenkins logs
virtually impossible.

Change-Id: Id64e6e1113fde08bfd476b0cfa3fc716f7b9801e