Update patch set 1
Patch Set 1:

> > Patch Set 1: Code-Review-2
 > 
 > Right, that's correct Kaspars. They're in a separate gate than the
 > unit tests.

I created this PS because the functional tests no longer seem to run correctly for me via `tox -e functional`.  Adding them to the gate was intended to be a sanity check for me to see whether my environment was somehow broken.

It appeared in my environment that the database tables were simply not created for the functional tests.

It appears to pass here when running after the unit tests.  This makes me suspect that the functional tests in the default tox job that I added are just using the tables setup by the unittests, and pushing  with the default env list in a separate order may cause them to fail.

I have no objection to the functional tests being a separate pipeline, but I still want to be able to run them without special setup in Jenkins.

I will change the order and push a new PS to test the hypothesis that it is just using the tables created by the unit tests.

Patch-set: 1
Label: Code-Review=0