Allow parentSelector to use multiple labels to select parent document
This PS updates the layering module so that a child document with
a compound parentSelector like: [{'a': 'b'}, {'c': 'd'}] can select
a parent with the exact same labels. Further, this works if
child's parentSelector is a sub-subset of parent's labels.

Adds unit tests for positive and negative test cases.

Change-Id: I7c4aac583365d90803eda77b82763decd78cfdcf