Update patch set 3
Patch Set 3:

Only tangentially related, I wonder if there is a way to warn on json paths that contain ":" characters in them, which are valid Values paths in helm but require quoting in json sources and destination paths.

e.g.

  .foo.bar.bing:baz.goo

  needs to be

  .foo.bar.'bing:baz'.goo

Today, Deckhand -- or really the JSON path parsing -- returns a parse error on ':'

Patch-set: 3