Run unit tests with sqlite
This PS makes unit tests run with sqlite rather than postgresql.
Unit tests can still be run with postgresql via:

    tox -e py35-postgresql

(as well as tox the py27 variety).

This also updates the remaining DB columns to use JSONB if
postgresql is enabled; else fallback columns are used for testing
with sqlite.

Change-Id: I951f2f04fd013d635bb7653a238ff1eb3725b5e1