Fix jsonpath_replace failing to create missing array keys
This PS resolves two issues, both related to not dynamically creating
missing arrays in the destination document, when the substitution
destination path indicates that an array should be created.

The first issue is that the array itself isn't created. For example,
if substitutions.dest.path = '.foo.bar[0].baz' then the data
section of the destination document should render like this:

   data:
     foo:
       bar:
         - baz: <some-value>

This PS also handles other edge cases like
substitutions.dest.path = '.foo.bar[3].baz' which should render as:

  data:
     foo:
       bar:
         - baz: <some-value>

Assuming that values for indices 0-2 aren't provided.

This PS also handles the case of nested arrays. Given
substitutions.dest.path = '.foo.bar[3].baz[0]', should render as:

  data:
     foo:
       bar:
         - baz:
           - <some-data>

As well as several other edge cases.

The second issue this PS resolves is not nuking the destination
document data if JSON path creation failed (for whatever reason).
This means that if jsonpath_replace in utils returns None,
then a warning is instead logged, with no substitution taking
place for a specific substitution in a document.

Change-Id: I87d0bb606b74fc1e05669da639ab22ec7bf55b25